PB503 MVC StudentWeb_Rauf rev1

+++ StudentWEB-rauf/Entity/Point.cs
@@ -6,7 +6,7 @@ namespace StudentWEB_rauf.Entity
     {
 
         public PointType PointType { get; set; }
-        public double Result { get; set; }
+        public decimal Result { get; set; }
         public int StudentId { get; set; }
         public Student Student { get; set; }
     }
